完成重构的显示阶段
