+++ app/src/main/java/com/binny/demo/MainActivity.java
@@ -6,11 +6,8 @@ import android.util.Log;
 import android.view.View;
 
 import com.binny.lib.DialogCalenderChoose;
-import com.binny.lib.bean.CalendarSelectResultBean;
-import com.binny.lib.callback.OnCalendarSelectResultCallback;
-import com.binny.lib.util.Logger;
 
-public class MainActivity extends AppCompatActivity implements OnCalendarSelectResultCallback {
+public class MainActivity extends AppCompatActivity {
     private final String TAG = this.getClass().getSimpleName();
 
     @Override
@@ -21,14 +18,8 @@ public class MainActivity extends AppCompatActivity implements OnCalendarSelectR
 
     public void show(View view) {
         new DialogCalenderChoose(this)
-                .setOnCalendarResult(this)
                 .setGregorianMonth(2015, 2029, 0).show();
     }
 
 
-    @Override
-    public void onSelectResult(CalendarSelectResultBean calendarSelectResultBean) {
-        Log.i(TAG, "onSelectResult: " + calendarSelectResultBean.getStartDay().getDayLongValue() + "  ---  " + calendarSelectResultBean.getEndDay().getDayLongValue());
-
-    }
 }
