+++ calendar/src/main/java/com/binny/lib/adapter/CalendarRVAdapter.java
@@ -1,6 +1,6 @@
 package com.binny.lib.adapter;
 
-import android.content.Context;
+import android.app.Activity;
 import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
 import android.util.Log;
@@ -12,12 +12,7 @@ import android.widget.TextView;
 
 import com.binny.lib.R;
 import com.binny.lib.bean.CalendarDateBean;
-import com.binny.lib.bean.MonthBean;
-import com.binny.lib.callback.OnFromToDateCallback;
-import com.binny.lib.util.Logger;
-import com.binny.lib.viewholder.monthholder.CalendarGVViewHolderHelper;
-import com.binny.lib.callback.OnCalendarSelectResultCallback;
-import com.smart.holder.CommonAdapter;
+import com.binny.lib.util.CalendarUtil;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,11 +25,23 @@ import static com.binny.lib.constant.CalendarConstant.PLACE_HOLDER;
  * describe
  */
 public class CalendarRVAdapter extends RecyclerView.Adapter<CalendarRVAdapter.ViewHolder> {
-    private Context mContext;
+    private Activity mActivity;
 
     private List<CalendarDateBean> mCalendarDateBeans = new ArrayList<>();//
-    public CalendarRVAdapter(Context context, List<CalendarDateBean> dateBeanList, OnCalendarSelectResultCallback selectResultCallback, OnFromToDateCallback onFromToDateCallback) {
-        mContext = context;
+    private int[] mClickedCount = new int[1];
+
+    public interface OnItemClickedListener {
+        void onItemClickedListener(TextView textView, CalendarDateBean.Day day, final int[] clickedCount);
+    }
+
+    private OnItemClickedListener mItemClickedListener;
+
+    public void setOnItemClickedListener(final OnItemClickedListener itemClickedListener) {
+        mItemClickedListener = itemClickedListener;
+    }
+
+    public CalendarRVAdapter(Activity activity, List<CalendarDateBean> dateBeanList) {
+        mActivity = activity;
         mCalendarDateBeans.clear();
         mCalendarDateBeans.addAll(dateBeanList);
     }
@@ -46,26 +53,54 @@ public class CalendarRVAdapter extends RecyclerView.Adapter<CalendarRVAdapter.Vi
     @Override
     public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         ViewHolder viewHolder;
-        View view = LayoutInflater.from(mContext).inflate(R.layout.layout_new_calendar_text_view_item, parent, false);
+        View view = LayoutInflater.from(mActivity).inflate(R.layout.layout_new_calendar_text_view_item, parent, false);
         viewHolder = new ViewHolder(view);
         return viewHolder;
     }
 
     @Override
-    public void onBindViewHolder(ViewHolder holder, int position) {
+    public void onBindViewHolder(final ViewHolder holder, int position) {
 
+        holder.setIsRecyclable(false);
         String title =mCalendarDateBeans.get(position).getMonthTitle();
         if (!PLACE_HOLDER.equals(title)) {
-            holder.mDay.setText(title);
-            holder.mDay.setPadding(20,5,5,5);
-            holder.mDay.setGravity(Gravity.CENTER_VERTICAL|Gravity.START);
+            holder.mDayTv.setText(title);
+            holder.mDayTv.setPadding(20, 5, 5, 5);
+            holder.mDayTv.setGravity(Gravity.CENTER_VERTICAL | Gravity.START);
         }else {
-            String string = mCalendarDateBeans.get(position).getDay().getDayInMonth();
-            holder.mDay.setText(string);
-            holder.mDay.setGravity(Gravity.CENTER);
+            final CalendarDateBean.Day day = mCalendarDateBeans.get(position).getDay();
+            if (day.isChosenStatus()) {
+                setCircle(holder.mDayTv);
+                Log.i("[holder, position]", "onBindViewHolder = " + day.getDayLongValue());
+            }
+            String string = day.getDayInMonth();
+            holder.mDayTv.setText(string);
+            holder.mDayTv.setGravity(Gravity.CENTER);
+            if (!TextUtils.isEmpty(string)) {
+                holder.mDayTv.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(final View v) {
+                        mItemClickedListener.onItemClickedListener(holder.mDayTv, day, mClickedCount);
+                    }
+                });
+            }
+
         }
     }
 
+    private void setRight(final TextView textView) {
+        textView.setBackground(mActivity.getResources().getDrawable(R.drawable.shape_calender_right));
+    }
+
+    /**
+     * @param textView 设置此view 的背景色
+     */
+    private void setCircle(final TextView textView) {
+        textView.setTextColor(mActivity.getResources().getColor(R.color.white));
+        textView.setBackground(mActivity.getResources().getDrawable(R.drawable.shape_calender_circle));
+        textView.setGravity(Gravity.CENTER);
+        CalendarUtil.changeWrapContent(textView);
+    }
     @Override
     public int getItemCount() {
         int size = mCalendarDateBeans.size();
@@ -76,11 +111,11 @@ public class CalendarRVAdapter extends RecyclerView.Adapter<CalendarRVAdapter.Vi
     }
 
     static class ViewHolder extends RecyclerView.ViewHolder {
-        private TextView mDay;
+        private TextView mDayTv;
 
         ViewHolder(View itemView) {
             super(itemView);
-            mDay = itemView.findViewById(R.id.calender_day_tv);
+            mDayTv = itemView.findViewById(R.id.calender_day_tv);
         }
     }
 }
