+++ calendar/src/main/java/com/binny/lib/bean/CalendarDateBean.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
  */
 public class CalendarDateBean implements Serializable {
     private String mMonthTitle;//月份
-    private MonthBean.Day mDay; //某一天
+    private Day mDay; //某一天
 
     public String getMonthTitle() {
         return mMonthTitle;
@@ -18,11 +18,128 @@ public class CalendarDateBean implements Serializable {
         mMonthTitle = monthTitle;
     }
 
-    public MonthBean.Day getDay() {
+    public Day getDay() {
         return mDay;
     }
 
-    public void setDay(final MonthBean.Day day) {
+    public void setDay(final Day day) {
         mDay = day;
+    }
+    /**
+     * 月份中的  某一天
+     */
+    public static class Day implements Serializable {
+        private  int mWeek;//周期
+        private String mDay;//日
+        private String mMonthTitle;//月
+        private String mYear;//年
+        private boolean mChosenStatus;//是否被选中
+        private int mDayLongValue;//用于比较的值
+        private boolean mStartPos;
+        private boolean mInitStatus;//为默选中值
+        private String mMonthInt;
+
+
+        public String getMonth() {
+            return mMonthTitle;
+        }
+
+        public void setMonth(String monthTitle) {
+            mMonthTitle = monthTitle;
+        }
+
+        public String getYear() {
+            return mYear;
+        }
+
+        public void setYear(String year) {
+            mYear = year;
+        }
+
+        public boolean isStartDay() {
+            return mIsFirstDay;
+        }
+
+        public void setFirstDay(boolean firstDay) {
+            mIsFirstDay = firstDay;
+        }
+
+        private boolean mIsFirstDay;
+
+        public boolean isEndDay() {
+            return mIsEndDay;
+        }
+
+        public void setEndDay(boolean endDay) {
+            mIsEndDay = endDay;
+        }
+
+        private boolean mIsEndDay;
+
+        public boolean isStartPos() {
+            return mStartPos;
+        }
+
+        public void setStartPos(boolean startPos) {
+            mStartPos = startPos;
+        }
+
+        public boolean isEndPos() {
+            return mEndPos;
+        }
+
+        public void setEndPos(boolean endPos) {
+            mEndPos = endPos;
+        }
+
+        private boolean mEndPos;
+
+        public int getDayLongValue() {
+            return mDayLongValue;
+        }
+
+        public void setDayLongValue(int longValue) {
+            mDayLongValue = longValue;
+        }
+
+        public  int getDayInWeek() {
+            return mWeek;
+        }
+
+        public void setDayInWeek(int week) {
+            mWeek = week;
+        }
+
+        public String getDayInMonth() {
+            return mDay;
+        }
+
+        public void setDayInMonth(String day) {
+            mDay = day;
+        }
+
+        public boolean isChosenStatus() {
+            return mChosenStatus;
+        }
+
+        public void setChosenStatus(boolean chosenStatus) {
+            mChosenStatus = chosenStatus;
+        }
+
+        public void setInitStatus(boolean initStatus) {
+            mInitStatus = initStatus;
+        }
+
+        public boolean isInitStatus() {
+            return mInitStatus;
+        }
+
+        public void setMonthInt(final String monthInt) {
+            mMonthInt = monthInt;
+        }
+
+        public String getMonthInt() {
+            return mMonthInt;
+        }
     }
 }
